+++ src/index.js
@@ -5,7 +5,7 @@ import { createMuiTheme, ThemeProvider } from "@material-ui/core/styles";
 
 import "./index.css";
 
-import App from "./components/App";
+import App from "./App";
 import Firebase, { FirebaseContext } from "./components/Firebase";
 
 const theme = createMuiTheme({
